remove old commit-verify
