+++ commit_verify/src/merkle.rs
@@ -222,3 +222,200 @@ where &'a T: CommitEncode
 
     fn merkle_leafs(&self) -> Self::LeafIter { self.iter() }
 }
+
+/*
+#[cfg(test)]
+mod test {
+    use std::collections::BTreeMap;
+
+    use amplify::{bmap, s};
+    use bitcoin_hashes::hex::ToHex;
+    use bitcoin_hashes::{sha256d, Hash};
+    use confined_encoding::{ConfinedDecode, ConfinedEncode};
+
+    use super::*;
+    use crate::commit_encode::{strategies, Strategy};
+    use crate::CommitConceal;
+
+    #[test]
+    fn collections() {
+        // First, we define a data type
+        #[derive(
+        Clone,
+        PartialEq,
+        Eq,
+        PartialOrd,
+        Ord,
+        Hash,
+        Debug,
+        ConfinedEncode,
+        ConfinedDecode
+        )]
+        struct Item(pub String);
+        // Next, we say that it should be concealed using some function
+        // (double SHA256 hash in this case)
+        impl CommitConceal for Item {
+            type ConcealedCommitment = sha256d::Hash;
+            fn commit_conceal(&self) -> Self::ConcealedCommitment {
+                sha256d::Hash::hash(self.0.as_bytes())
+            }
+        }
+        // Next, we need to specify how the concealed data should be
+        // commit-encoded: this time we strict-serialize the hash
+        impl Strategy for sha256d::Hash {
+            type Strategy = strategies::UsingStrict;
+        }
+        // Now, we define commitment encoding for our concealable type: it
+        // should conceal the data
+        impl Strategy for Item {
+            type Strategy = strategies::UsingConceal;
+        }
+        // Now, we need to say that consensus commit procedure should produce
+        // a final commitment from commit-encoded data (equal to the
+        // strict encoding of the conceal result) using `CommitVerify` type.
+        // Here, we use another round of hashing, producing merkle node hash
+        // from the concealed data.
+        impl ConsensusCommit for Item {
+            type Commitment = MerkleNode;
+        }
+        // Next, we need to provide merkle node tags for each type of the tree
+        impl ConsensusMerkleCommit for Item {
+            const MERKLE_NODE_PREFIX: &'static str = "item";
+        }
+        impl ConsensusMerkleCommit for (usize, Item) {
+            const MERKLE_NODE_PREFIX: &'static str = "usize->item";
+        }
+
+        impl ToMerkleSource for BTreeMap<usize, Item> {
+            type Leaf = (usize, Item);
+            fn to_merkle_source(&self) -> MerkleSource<Self::Leaf> {
+                self.iter().map(|(k, v)| (*k, v.clone())).collect()
+            }
+        }
+
+        let large = vec![Item(s!("none")); 3];
+        let vec: MerkleSource<Item> = large.into();
+        assert_eq!(
+            vec.commit_serialize().to_hex(),
+            "71ea45868fbd924061c4deb84f37ed82b0ac808de12aa7659afda7d9303e7a71"
+        );
+
+        let large = vec![Item(s!("none")); 5];
+        let vec: MerkleSource<Item> = large.into();
+        assert_eq!(
+            vec.commit_serialize().to_hex(),
+            "e255e0124efe0555fde0d932a0bc0042614129e1a02f7b8c0bf608b81af3eb94"
+        );
+
+        let large = vec![Item(s!("none")); 9];
+        let vec: MerkleSource<Item> = large.into();
+        assert_eq!(
+            vec.commit_serialize().to_hex(),
+            "6cd2d5345a654af4720bdcc637183ded8e432dc88f778b7d27c8d5a0e342c65f"
+        );
+
+        let large = vec![Item(s!("none")); 13];
+        let vec: MerkleSource<Item> = large.into();
+        assert_eq!(
+            vec.commit_serialize().to_hex(),
+            "3714c08c7c94a4ef769ad2cb7df9aaca1e1252d6599a02aff281c37e7242797d"
+        );
+
+        let large = vec![Item(s!("none")); 17];
+        let vec: MerkleSource<Item> = large.into();
+        assert_eq!(
+            vec.commit_serialize().to_hex(),
+            "6093dec47e5bdd706da01e4479cb65632eac426eb59c8c28c4e6c199438c8b6f"
+        );
+
+        let item = Item(s!("Some text"));
+        assert_eq!(
+            &b"\x09\x00Some text"[..],
+            item.confined_serialize().unwrap()
+        );
+        assert_eq!(
+            "6680bbec0d05d3eaac9c8b658c40f28d2f0cb0f245c7b1cabf5a61c35bd03d8e",
+            item.commit_serialize().to_hex()
+        );
+        assert_eq!(
+            "3e4b2dcf9bca33400028c8947565c1ff421f6d561e9ec48f88f0c9a24ebc8c30",
+            item.consensus_commit().to_hex()
+        );
+        assert_ne!(item.commit_serialize(), item.confined_serialize().unwrap());
+        assert_eq!(
+            MerkleNode::hash(&item.commit_serialize()),
+            item.consensus_commit()
+        );
+
+        let original = bmap! {
+            0usize => Item(s!("My first case")),
+            1usize => Item(s!("My second case with a very long string")),
+            3usize => Item(s!("My third case to make the Merkle tree two layered"))
+        };
+        let collection = original.to_merkle_source();
+        assert_eq!(
+            &b"\x03\x00\
+             \x00\x00\
+             \x0d\x00\
+             My first case\
+             \x01\x00\
+             \x26\x00\
+             My second case with a very long string\
+             \x03\x00\
+             \x31\x00\
+             My third case to make the Merkle tree two layered"[..],
+            original.confined_serialize().unwrap()
+        );
+        assert_eq!(
+            "d911717b8dfbbcef68495c93c0a5e69df618f5dcc194d69e80b6fafbfcd6ed5d",
+            collection.commit_serialize().to_hex()
+        );
+        assert_eq!(
+            "d911717b8dfbbcef68495c93c0a5e69df618f5dcc194d69e80b6fafbfcd6ed5d",
+            collection.consensus_commit().to_hex()
+        );
+        assert_ne!(
+            collection.commit_serialize(),
+            original.confined_serialize().unwrap()
+        );
+        assert_eq!(
+            MerkleNode::from_slice(&collection.commit_serialize()).unwrap(),
+            collection.consensus_commit()
+        );
+
+        let original = vec![
+            Item(s!("My first case")),
+            Item(s!("My second case with a very long string")),
+            Item(s!("My third case to make the Merkle tree two layered")),
+        ];
+        let vec: MerkleSource<Item> = original.clone().into();
+        assert_eq!(
+            &b"\x03\x00\
+             \x0d\x00\
+             My first case\
+             \x26\x00\
+             My second case with a very long string\
+             \x31\x00\
+             My third case to make the Merkle tree two layered"[..],
+            original.confined_serialize().unwrap()
+        );
+        assert_eq!(
+            "fd72061e26055fb907aa512a591b4291e739f15198eb72027c4dd6506f14f469",
+            vec.commit_serialize().to_hex()
+        );
+        assert_eq!(
+            "fd72061e26055fb907aa512a591b4291e739f15198eb72027c4dd6506f14f469",
+            vec.consensus_commit().to_hex()
+        );
+        assert_ne!(
+            vec.commit_serialize(),
+            original.confined_serialize().unwrap()
+        );
+        assert_eq!(
+            MerkleNode::from_slice(&vec.commit_serialize()).unwrap(),
+            vec.consensus_commit()
+        );
+        assert_ne!(vec.consensus_commit(), collection.consensus_commit());
+    }
+}
+*/
